ERICKsky-signal-engine/signal-bot/requirements.txt
# ERICKsky Signal Engine - Python Dependencies
# Python 3.11+

# Core async / task queue
celery==5.3.6
redis==5.0.1
kombu==5.3.4

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Market data
twelvedata==1.2.11
websocket-client==1.7.0
requests==2.31.0

# Data analysis
pandas==2.1.4
numpy==1.26.2

# Scheduling
APScheduler==3.10.4

# Telegram
python-telegram-bot==20.7
httpx==0.25.2

# Config & environment
python-dotenv==1.0.0

# Utilities
pytz==2024.1
python-dateutil==2.8.2
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
